+++ unix/tar_lwt_unix.ml
@@ -15,7 +15,7 @@
  * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  *)
 
-open Lwt
+open Lwt.Infix
 
 module Reader = struct
   type in_channel = Lwt_unix.file_descr
@@ -25,7 +25,7 @@ module Reader = struct
     let buffer_size = 32768 in
     let buffer = Cstruct.create buffer_size in
     let rec loop (n: int) =
-      if n <= 0 then return ()
+      if n <= 0 then Lwt.return_unit
       else
         let amount = min n buffer_size in
         let block = Cstruct.sub buffer 0 amount in
@@ -45,7 +45,7 @@ let copy_n ifd ofd n =
   let block_size = 32768 in
   let buffer = Cstruct.create block_size in
   let rec loop remaining =
-    if remaining = 0L then return () else begin
+    if remaining = 0L then Lwt.return_unit else begin
       let this = Int64.(to_int (min (of_int block_size) remaining)) in
       let block = Cstruct.sub buffer 0 this in
       really_read ifd block >>= fun () ->
@@ -60,8 +60,8 @@ module HW = Tar.HeaderWriter(Lwt)(Writer)
 let get_next_header ?level ic =
   HR.read ?level ic
   >>= function
-  | Error `Eof -> return None
-  | Ok hdr -> return (Some hdr)
+  | Error `Eof -> Lwt.return None
+  | Ok hdr -> Lwt.return (Some hdr)
 
 (** Return the header needed for a particular file on disk *)
 let header_of_file ?level (file: string) : Tar.Header.t Lwt.t =
@@ -83,8 +83,8 @@ let header_of_file ?level (file: string) : Tar.Header.t Lwt.t =
   Lwt.return (Tar.Header.make ~file_mode ~user_id ~group_id ~mod_time ~link_indicator ~link_name
     ~uname ~gname ~devmajor ~devminor file file_size)
 
-let write_block (header: Tar.Header.t) (body: Lwt_unix.file_descr -> unit Lwt.t) (fd : Lwt_unix.file_descr) =
-  HW.write header fd
+let write_block ?level (header: Tar.Header.t) (body: Lwt_unix.file_descr -> unit Lwt.t) (fd : Lwt_unix.file_descr) =
+  HW.write ?level header fd
   >>= fun () ->
   body fd >>= fun () ->
   really_write fd (Tar.Header.zero_padding header)
@@ -105,14 +105,14 @@ module Archive = struct
     | Some hdr ->
       f fd hdr >>= fun result ->
       Reader.skip fd (Tar.Header.compute_zero_padding_length hdr) >>= fun () ->
-      return (Some result)
+      Lwt.return (Some result)
     | None ->
-      return None
+      Lwt.return None
 
   (** List the contents of a tar *)
   let list ?level fd =
     let rec loop acc = get_next_header ?level fd >>= function
-      | None -> return (List.rev acc)
+      | None -> Lwt.return (List.rev acc)
       | Some hdr ->
         Reader.skip fd (Int64.to_int hdr.Tar.Header.file_size) >>= fun () ->
         Reader.skip fd (Tar.Header.compute_zero_padding_length hdr) >>= fun () ->
@@ -122,28 +122,40 @@ module Archive = struct
   (** Extract the contents of a tar to directory 'dest' *)
   let extract dest ifd =
     let rec loop () = get_next_header ifd >>= function
-      | None -> return ()
+      | None -> Lwt.return_unit
       | Some hdr ->
         let filename = dest hdr.Tar.Header.file_name in
-        Lwt_unix.openfile filename [Unix.O_WRONLY] 0644 >>= fun ofd ->
+        Lwt_unix.openfile filename [Unix.O_WRONLY] 0 >>= fun ofd ->
         copy_n ifd ofd hdr.Tar.Header.file_size >>= fun () ->
         Reader.skip ifd (Tar.Header.compute_zero_padding_length hdr) >>= fun () ->
         loop () in
     loop ()
+
+  let transform ?level f ifd ofd =
+    let rec loop () = get_next_header ifd >>= function
+      | None -> Lwt.return_unit
+      | Some header' ->
+        let header = f header' in
+        let body = fun _ -> copy_n ifd ofd header.Tar.Header.file_size in
+        write_block ?level header body ofd >>= fun () ->
+        Reader.skip ifd (Tar.Header.compute_zero_padding_length header') >>= fun () ->
+        loop () in
+    loop () >>= fun () ->
+    write_end ofd
 
   (** Create a tar on file descriptor fd from the filename list
       'files' *)
   let create files ofd =
     let file filename =
       Lwt_unix.stat filename >>= fun stat ->
-      if stat.Unix.st_kind <> Unix.S_REG then begin
-        Printf.eprintf "Skipping %s: not a regular file\n" filename;
-        return ()
-      end else begin
+      if stat.Unix.st_kind <> Unix.S_REG then
+        (* Skipping, not a regular file. *)
+        Lwt.return_unit
+      else begin
         header_of_file filename >>= fun hdr ->
 
         write_block hdr (fun ofd ->
-            Lwt_unix.openfile filename [Unix.O_RDONLY] 0644 >>= fun ifd ->
+            Lwt_unix.openfile filename [Unix.O_RDONLY] 0 >>= fun ifd ->
             copy_n ifd ofd hdr.Tar.Header.file_size
           ) ofd
       end in