+++ CHANGES.md
@@ -2,7 +2,8 @@
 
 - Bump lower-bound on Cstruct to 6.0.0 (@MisterDA)
 - Update to Dune 2.9 and generate opam files (@MisterDA)
-- Don't print the name of the file in extract functions (@MisterDA)
+- Support only OCaml versions 4.08 and higher. (@MisterDA)
+- Don't print any logging to stdout or stderr (@MisterDA)
 - Remove Tar.Make.Header, Tar_cstruct.Header, Tar_unix.Header, and
   Tar_lwt_unix.Header to keep only Tar.Header and use it everywhere.
   - Tar.Make.Header.get_next_header becomes Tar.Make.get_next_header;
@@ -13,6 +14,11 @@
   - Tar_unix.Header.of_file becomes Tar_unix.header_of_file;
   - All the Tar_*.Header.t values have to be changed to Tar.Header.t.
   (@MisterDA)
+- Fix parsing of pax Extended Header File Times with sub-second
+  granularity. (@MisterDA)
+- Add Tar_unix.transform and Tar_lwt_unix.transform to help
+  transforming headers of a streamed tar archive between two file
+  descriptors. (@MisterDA)
 
 ## v1.1.0 (2019-04-08)